pyproject.toml
[project]
name = "2nd-structure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "pymolpy3>=0.1.2",
]
